test(test-utils): promote SSE helpers from zfb-server integration tests (#1018)
Move `next_sse_event_name` + `wait_for_subscribers` from
crates/zfb-server/tests/integration.rs into zfb-test-utils so the new
`zfb dev` E2E harness (crates/zfb/tests/dev_serve_e2e.rs) can reuse
them. Mechanical promotion; the only change is making
`wait_for_subscribers` generic over the broadcast payload type so
zfb-test-utils needs no zfb-server dependency (which would have created
a dev-dep cycle). integration.rs now imports both from the shared crate.
