+++ crates/zfb-server/Cargo.toml
@@ -72,3 +72,6 @@ tokio = { version = "1", features = ["sync", "rt", "macros", "time", "rt-multi-t
 tempfile = { workspace = true }
 reqwest = { version = "0.12", default-features = false, features = ["stream"] }
 futures-util = "0.3"
+# Shared SSE test helpers (next_sse_event_name / wait_for_subscribers),
+# promoted from this crate's integration.rs in #1018.
+zfb-test-utils = { path = "../zfb-test-utils" }
